converter, storage: extract disk conversion logic
Move Convert_v1_Disk_To_api_Disk, assignDiskToSCSIController,
setReservation, SetErrorPolicy, ToApiReadOnly, AssignDiskIOThread,
and DeviceTypeNotCompatibleFmt from converter.go to storage/disk.go.
