+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "blockio.go",
         "devicenaming.go",
         "directio.go",
+        "disk.go",
         "driver.go",
         "generated_mock_directio.go",
         "paths.go",
@@ -17,8 +18,14 @@ go_library(
         "//pkg/os/disk:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/safepath:go_default_library",
+        "//pkg/storage/reservation:go_default_library",
+        "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/iothreads:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/virtio:go_default_library",
+        "//pkg/virt-launcher/virtwrap/device:go_default_library",
         "//pkg/virt-launcher/virtwrap/disksource:go_default_library",
         "//pkg/virtiofs:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
@@ -34,6 +41,7 @@ go_test(
         "blockio_test.go",
         "devicenaming_test.go",
         "directio_test.go",
+        "disk_test.go",
         "driver_test.go",
         "storage_suite_test.go",
         "virtiofs_test.go",
@@ -44,10 +52,14 @@ go_test(
         "//pkg/libvmi:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/arch:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/go.uber.org/mock/gomock:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
     ],
 )
