+++ pkg/virt-launcher/virtwrap/converter/storage/disk.go
@@ -1,0 +1,191 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright The KubeVirt Authors.
+ *
+ */
+
+package storage
+
+import (
+	"fmt"
+	"slices"
+	"strconv"
+
+	v1 "kubevirt.io/api/core/v1"
+
+	"kubevirt.io/kubevirt/pkg/pointer"
+	"kubevirt.io/kubevirt/pkg/storage/reservation"
+	storagetypes "kubevirt.io/kubevirt/pkg/storage/types"
+	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
+	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/iothreads"
+	convertertypes "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/types"
+	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/virtio"
+	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device"
+)
+
+const (
+	DeviceTypeNotCompatibleFmt = "device %s is of type lun. Not compatible with a file based disk"
+)
+
+func Convert_v1_Disk_To_api_Disk(c *convertertypes.ConverterContext, diskDevice *v1.Disk, disk *api.Disk, prefixMap map[string]DeviceNamer, numQueues *uint, volumeStatusMap map[string]v1.VolumeStatus) error {
+	if diskDevice.Disk != nil {
+		var unit int
+		disk.Device = "disk"
+		disk.Target.Bus = diskDevice.Disk.Bus
+		disk.Target.Device, unit = MakeDeviceName(diskDevice.Name, diskDevice.Disk.Bus, prefixMap)
+		if diskDevice.Disk.Bus == "scsi" {
+			assignDiskToSCSIController(disk, unit)
+		}
+		if diskDevice.Disk.PciAddress != "" {
+			if diskDevice.Disk.Bus != v1.DiskBusVirtio {
+				return fmt.Errorf("setting a pci address is not allowed for non-virtio bus types, for disk %s", diskDevice.Name)
+			}
+			addr, err := device.NewPciAddressField(diskDevice.Disk.PciAddress)
+			if err != nil {
+				return fmt.Errorf("failed to configure disk %s: %v", diskDevice.Name, err)
+			}
+			disk.Address = addr
+		}
+		if diskDevice.Disk.Bus == v1.DiskBusVirtio {
+			disk.Model = virtio.InterpretTransitionalModelType(&c.UseVirtioTransitional, c.Architecture.GetArchitecture())
+		}
+		disk.ReadOnly = ToApiReadOnly(diskDevice.Disk.ReadOnly)
+		disk.Serial = diskDevice.Serial
+		if diskDevice.Shareable != nil {
+			if *diskDevice.Shareable {
+				if diskDevice.Cache == "" {
+					diskDevice.Cache = v1.CacheNone
+				}
+				if diskDevice.Cache != v1.CacheNone {
+					return fmt.Errorf("a sharable disk requires cache = none got: %v", diskDevice.Cache)
+				}
+				disk.Shareable = &api.Shareable{}
+			}
+		}
+	} else if diskDevice.LUN != nil {
+		var unit int
+		disk.Device = "lun"
+		disk.Target.Bus = diskDevice.LUN.Bus
+		disk.Target.Device, unit = MakeDeviceName(diskDevice.Name, diskDevice.LUN.Bus, prefixMap)
+		if diskDevice.LUN.Bus == "scsi" {
+			assignDiskToSCSIController(disk, unit)
+		}
+		disk.ReadOnly = ToApiReadOnly(diskDevice.LUN.ReadOnly)
+		if diskDevice.LUN.Reservation {
+			setReservation(disk)
+		}
+	} else if diskDevice.CDRom != nil {
+		disk.Device = "cdrom"
+		disk.Target.Tray = string(diskDevice.CDRom.Tray)
+		disk.Target.Bus = diskDevice.CDRom.Bus
+		disk.Target.Device, _ = MakeDeviceName(diskDevice.Name, diskDevice.CDRom.Bus, prefixMap)
+		if diskDevice.CDRom.ReadOnly != nil {
+			disk.ReadOnly = ToApiReadOnly(*diskDevice.CDRom.ReadOnly)
+		} else {
+			disk.ReadOnly = ToApiReadOnly(true)
+		}
+	}
+	disk.Driver = &api.DiskDriver{
+		Name:  "qemu",
+		Cache: string(diskDevice.Cache),
+		IO:    diskDevice.IO,
+	}
+	if diskDevice.Disk != nil || diskDevice.LUN != nil {
+		if !slices.Contains(c.VolumesDiscardIgnore, diskDevice.Name) {
+			disk.Driver.Discard = "unmap"
+		}
+		volumeStatus, ok := volumeStatusMap[diskDevice.Name]
+		if ok && volumeStatus.PersistentVolumeClaimInfo != nil {
+			disk.FilesystemOverhead = volumeStatus.PersistentVolumeClaimInfo.FilesystemOverhead
+			disk.Capacity = storagetypes.GetDiskCapacity(volumeStatus.PersistentVolumeClaimInfo)
+		}
+	}
+	if numQueues != nil && disk.Target.Bus == v1.DiskBusVirtio {
+		disk.Driver.Queues = numQueues
+	}
+	disk.Alias = api.NewUserDefinedAlias(diskDevice.Name)
+	if diskDevice.BootOrder != nil {
+		disk.BootOrder = &api.BootOrder{Order: *diskDevice.BootOrder}
+	}
+	if (c.UseLaunchSecuritySEV || c.UseLaunchSecurityPV) && disk.Target.Bus == v1.DiskBusVirtio {
+		disk.Driver.IOMMU = "on"
+	}
+
+	return nil
+}
+
+func AssignDiskIOThread(disk *v1.Disk, apiDisk *api.Disk, supplementalIOThreads *api.DiskIOThreads, autoThreads int, currentDedicatedThread, currentAutoThread uint) (uint, uint) {
+	if apiDisk.Target.Bus == v1.DiskBusVirtio {
+		if supplementalIOThreads != nil {
+			apiDisk.Driver.IOThreads = supplementalIOThreads
+		} else {
+			if iothreads.HasDedicatedIOThread(*disk) {
+				apiDisk.Driver.IOThread = pointer.P(currentDedicatedThread)
+				currentDedicatedThread += 1
+			} else {
+				apiDisk.Driver.IOThread = pointer.P(currentAutoThread)
+				// increment the threadId to be used next but wrap around at the thread limit
+				// the odd math here is because thread ID's start at 1, not 0
+				currentAutoThread = (currentAutoThread % uint(autoThreads)) + 1
+			}
+		}
+	}
+	return currentDedicatedThread, currentAutoThread
+}
+
+func ToApiReadOnly(src bool) *api.ReadOnly {
+	if src {
+		return &api.ReadOnly{}
+	}
+	return nil
+}
+
+func SetErrorPolicy(diskDevice *v1.Disk, disk *api.Disk) error {
+	if diskDevice.ErrorPolicy == nil {
+		disk.Driver.ErrorPolicy = v1.DiskErrorPolicyStop
+		return nil
+	}
+	switch *diskDevice.ErrorPolicy {
+	case v1.DiskErrorPolicyStop, v1.DiskErrorPolicyIgnore, v1.DiskErrorPolicyReport, v1.DiskErrorPolicyEnospace:
+		disk.Driver.ErrorPolicy = *diskDevice.ErrorPolicy
+	default:
+		return fmt.Errorf("error policy %s not recognized", *diskDevice.ErrorPolicy)
+	}
+	return nil
+}
+
+func assignDiskToSCSIController(disk *api.Disk, unit int) {
+	// Ensure we assign this disk to the correct scsi controller
+	if disk.Address == nil {
+		disk.Address = &api.Address{}
+	}
+	disk.Address.Type = "drive"
+	// This should be the index of the virtio-scsi controller, which is hard coded to 0
+	disk.Address.Controller = "0"
+	disk.Address.Bus = "0"
+	disk.Address.Unit = strconv.Itoa(unit)
+}
+
+func setReservation(disk *api.Disk) {
+	disk.Source.Reservations = &api.Reservations{
+		Managed: "no",
+		SourceReservations: &api.SourceReservations{
+			Type: "unix",
+			Path: reservation.GetPrHelperSocketPath(),
+			Mode: "client",
+		},
+	}
+}
